Set up Jest for content editor testing
