+++ jest.config.js
@@ -1,0 +1,6 @@
+import { createJestConfig } from '@craco/craco';
+
+import cracoConfig from './craco.config.js';
+const jestConfig = createJestConfig(cracoConfig);
+
+export default jestConfig;
