+++ package.json
@@ -37,13 +37,14 @@
     "reactstrap": "^8.10.1",
     "remarkable": "^2.0.1",
     "swr": "^2.0.3",
+    "ts-jest": "^29.2.5",
     "typescript": "^4.5.5",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "craco start --verbose",
     "build": "craco build",
-    "test": "craco test",
+    "test": "craco test --watchAll",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
